requirements.txt
Flask==3.1.1
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.3
psycopg2-binary==2.9.10